pyproject.toml
[tool.poetry]
name = "blog_server"
version = "0.1.0"
description = ""
authors = ["tariel"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0.7"
djangorestframework = "^3.11.0"
psycopg2-binary = "^2.8.5"
python-language-server = "^0.33.1"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
